+++ scripts/shell/2-make-box-from-packer.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
 # Simple packer build, vbox only
-packer build -only=virtualbox-iso template.json
+packer build template.json